Create local/; move local python impl.
We're creating a web app version of this so the python implmentation is
no longer necessary. We could have written it so the python
implementation kept a local copy of the repository content but I don't
know of a web service that allows us to store so much data so instead
we'll get the content from hgweb.